- Squash two other compile warnings. ircd now compiles clean with   -Wall   -Wextra   -Wno-unused   -Wcast-qual   -Wcast-align   -Wbad-function-cast   -Wmissing-declarations   -Wmissing-prototypes   -Wnested-externs   -Wredundant-decls   -Wshadow   -Wwrite-strings   -Wundef
git-svn-id: svn://svn.ircd-hybrid.org/svnroot/ircd-hybrid-8@1330 82007160-df01-0410-b94d-b575c5fd34c7

+++ src/fdlist.c
@@ -49,7 +49,7 @@ changing_fdlimit(va_list args)
 
   hard_fdlimit = va_arg(args, int);
 
-  if (ServerInfo.max_clients > MAXCLIENTS_MAX)
+  if (ServerInfo.max_clients > (unsigned int)MAXCLIENTS_MAX)
   {
     if (old_fdlimit != 0)
       sendto_realops_flags(UMODE_ALL, L_ALL,
